Cleanup and optimise the Kustomization resources
Cleanup, standardise, and optimise the Kustomization resources for the
two clusters, with a standard layout and common settings. Also increase
the interval for reconciliation from 1m to 1h as new changes are
immediate, and there isn't the need to check for configuration drive
that often.

Also configure the namespace explicitly for the source GitRepository,
and override the interview for failures to allow it to retry more often
until a successful result.