Add new permissions Kustomization for all Clusters
Add a new permissions Kustomization to be applied to all Clusters which
will provide standard additional permissions, as needed. Specifically
this is for GitOps currently as provides a link between the username
(i.e. email address) from Auth0 which is used to log into GitOps and the
permissions of that username on the cluster itself. Without this
permission the interface is empty as the user cannot query the
resources.